Update veteran date of death with correct date format (#16180)
### Description
This PR fixes a bug in `Veteran#update_cached_attributes!` which incorrectly interprets the format of date of death from BGS.

The fix was cherry-picked from @bgantick's work on #16019 which provides a more holistic approach to dealing with the date of death caching.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
Added new rspec for `Veteran#update_cached_attributes!`

When this test is run with the old version of `update_cached_attributes!`, the failure diff is given as

```
Veteran#update_cached_attributes! when date of death is present saves date of death in the correct date format
Failure/Error: expect(veteran[:date_of_death]).to eq(new_date_of_death)

  expected: Mon, 08 Mar 2021 
       got: Tue, 03 Aug 2021 

  (compared using ==)
```

which matches the data error we were seeing in the wild.