Only show slots for virtual hearing scheduling (#16150)
[CASEFLOW-1115](https://vajira.max.gov/browse/CASEFLOW-1115) will be done across multiple PRs:
- Backend changes ([16147](https://github.com/department-of-veterans-affairs/caseflow/pull/16147)), adds db columns, falls back to existing behavior for number_of_slots and begins_at (slot_length_minutes has no existing behavior in db).
- Only show slots when scheduling virtual hearings ([16150](https://github.com/department-of-veterans-affairs/caseflow/pull/16150)).
- UI Fixes PR ([16149](https://github.com/department-of-veterans-affairs/caseflow/pull/16149)), stop using HearingTime to fix various bugs (temporary fix until we replace with custom time modal).
- Slot recalculation PR ([16161](https://github.com/department-of-veterans-affairs/caseflow/pull/16161)), branch is based on UI Fixes PR, accept new db values and calculate slots based on them.

### Description
This is a tiny PR that makes it so slots will only show up if the user is scheduling a virtual hearing.

This doesn't quite match the language in JIRA, but I believe is the correct solutions.
```
Display the time slots only on hearing days with a type of "Virtual"
i.e., "Video" and "Central" keep using the radio buttons and dropdown
```

### Testing Plan
Verify that when scheduling the slots only appear for virtual hearings. You can do this easily by changing the type of the hearing in the scheduling screen from "Video" to "Virtual" and back.