Show substitution granted date; adds AppellantSubstitution factory (#16244)
Resolves [CASEFLOW-1437](https://vajira.max.gov/browse/CASEFLOW-1437)

### Description
When a substitute appellant is present, the date the substitution was granted by the RO is now shown.

In doing so, I added a simple AppellantSubstitution factory and have utils.js load the AppellantSubstitution that the serializer was already making available.

### Acceptance Criteria
- [x] The page displays normally when there _isn't_ a substitute appellant (i.e., any random appeal you pick still renders)
- [x] Viewing an appeal with a target appellant (see testing plan below) shows substitution date
- [x] UI component matches [the Figma design](https://www.figma.com/file/de0tCLh5ITud8b8QuII0Hh/Granted-Substitutions-2021?node-id=1266%3A2006).

### Testing Plan

You may need to clear out existing front-end stuff with `make clean` first.

Create a new AppellantSubstitution:

```ruby
as = FactoryBot.create(:appellant_substitution)
new_appeal = as.target_appeal.uuid
```

Visit `new_appeal` in your browser.

Under "About the Appellant," you should see a new "Substitution granted by the RO" section. It's probably a good idea to confirm that the date matches the one on the `as` object, and that it's displayed in a more human-friendly way in the UI.

Confirm that any other random appeal in the UI doesn't show this (unless you created others with substitute appellants), and doesn't display a fatal error. (Ahem, not that I had temporarily broken that or anything. ;))

### User Facing Changes
 - [x] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER
---|---
![Screen Shot 2021-05-14 at 4 40 42 PM](https://user-images.githubusercontent.com/314663/118326583-27935c00-b4d3-11eb-8edf-0cca13cf6f55.png) | ![Screen Shot 2021-05-14 at 4 33 14 PM](https://user-images.githubusercontent.com/314663/118325955-3dece800-b4d2-11eb-8edd-807183d383de.png)