Add front end validation to receipt date field (#16280)
Resolves CASEFLOW-1566

### Description
Add validation to the receipt date field to ensure that the date entered is not prior to the AMA start date

This should be considered a quick fix to prevent any more intakes from being entered with an early receipt date, since we're just using vanilla yup to validate the data.

A good IP ticket for this coming IP sprint would be to refactor the way these forms work, since they're not actually html forms so there's not a real "submission" that we could use the more robust yupResolver to enhance the validations here. The refactor would include replacing the Class based components with functional components which would allow us to use form hooks. 

### Acceptance Criteria
- If a user enters a date prior to the AMA start date, chooses unrecognized claimant, and tries to move to next step, they should receive an error informing them that the receipt date cannot be before the AMA date

### Testing Plan
1. Start an NOD appeal
2. On the review page, enter a receipt date prior to 02/19/2019
3. Choose Yes for Claimant other than someone listed
4. Choose claimant not listed
5. Click continue to next step
6. See that there is an error on the receipt date field

- [ ] For higher-risk changes: [Deploy the custom branch to UAT to test](https://github.com/department-of-veterans-affairs/appeals-deployment/wiki/Applications---Deploy-Custom-Branch-to-UAT)

### User Facing Changes
 - [ ] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER
 ---|---

### Code Documentation Updates
- [ ] Add or update code comments at the top of the class, module, and/or component.

### Storybook Story
*For Frontend (Presentationa) Components*
* [ ] Add a [Storybook](https://github.com/department-of-veterans-affairs/caseflow/wiki/Documenting-React-Components-with-Storybook) file alongside the component file (e.g. create `MyComponent.stories.js` alongside `MyComponent.jsx`)
* [ ] Give it a title that reflects the component's location within the overall Caseflow hierarchy
* [ ] Write a separate story (within the same file) for each discrete variation of the component