Prevent Multiple Substitutions from Source Appeal (#16386)
Connects [CASEFLOW-1660](https://vajira.max.gov/browse/CASEFLOW-1660)

### Description
This adds data to `AppealSerializer` that shows substitutions performed for the source appeal. This gives us the ability to use logic on frontend to place informational alerts, etc if desired. In the scope of this ticket, it allows us to prevent the "Add Substitute" button from displaying if a substitution has already occurred on the given source appeal. 

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] "Add Substitute" button does not display if a substitution has already taken place on that appeal

### Testing Plan
1. Perform an appellant substitution (for instance using case for `54545454` veteran file number)
2. Go back to source appeal (via "View all cases" link or otherwise)
3. Verify that "Add Substitution" button no longer displays

### User Facing Changes
 - [ ] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER
 ---|---
![Screen Shot 2021-06-22 at 13 05 48](https://user-images.githubusercontent.com/2581274/122973417-346f5f00-d35f-11eb-850d-7c4c56d5218b.png) | ![Screen Shot 2021-06-22 at 13 05 58](https://user-images.githubusercontent.com/2581274/122973455-3df8c700-d35f-11eb-8c6e-fb5a6d6d5a45.png)