POA not found controller error, refactoring (#16390)
### Description
For the POA refresh button on the Case Details page, in cases where there was previously no POA, the button was sending a nil poaId to the request on the backend. This searched for the BgsPowerOfAttorney using the ID, and if it was not found, causing a 404 error on the request, most likely when attempting to call a method on it.

This PR fixes that bug and does some refactoring to simplify the logic in the Appeals Controller. It also removes connecting an Attorney Claimant's POA to themselves, which could give them extra access to their own case. Currently, POAs are not supported for Attorney Claimants.

### Acceptance Criteria
- [ ] POA update button succeeds when there is originally no POA, and still isn't one
- [ ] AttorneyClaimants return no POA

### Testing Plan
1. Feature test for the new scenario, "not found"

### User Facing Changes
 - [ ] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER
 ---|---
<img width="780" alt="Screen Shot 2021-06-23 at 6 32 22 PM" src="https://user-images.githubusercontent.com/6589610/123194941-294e2900-d45c-11eb-9c3b-a111d8d04ad3.png"> | <img width="783" alt="Screen Shot 2021-06-23 at 6 32 30 PM" src="https://user-images.githubusercontent.com/6589610/123194957-30753700-d45c-11eb-8ef5-809ba9d992c0.png">


### Code Documentation Updates
- [ ] Add or update code comments at the top of the class, module, and/or component.