Do not raise InvalidCountryNameError for nil appellant_address country (#16403)
Resolves the remaining [2 of 3 common `TimezoneService::InvalidCountryNameError`s](https://sentry.prod.appeals.va.gov/department-of-veterans-affairs/caseflow/issues/18565/) we have been seeing lately that were not address by the previous PR (#16399): 
```
TimezoneService::InvalidCountryNameError: no ISO 3166 country code found for ""
```

This PR handles the two scenarios in different ways:

### 1️⃣ Addresses with all `nil` fields

Some `LegacyAppeal`s (for example, 1357514 and 1357515) have an `appellant_address` with all fields present but with `nil` values:
```ruby
{
  :address_line_1=>nil,
  :address_line_2=>nil,
  :address_line_3=>nil,
  :city=>nil,
  :state=>nil,
  :country=>nil,
  :zip=>nil
}
``` 

Behaviour before this change | After
--- | ---
As written, the code creates an `Address` object for addresses like these and calls `TimezoneService.address_to_timezone()` which [fails when trying to determine the country](https://github.com/department-of-veterans-affairs/caseflow/blob/8e547951af3d4ff727fac940aead741f55f5b796/app/services/timezone_service.rb#L30) from the input `Address` and sends an `InvalidCountryNameError` above to Sentry and returns a `nil` value to the serializer. | With this change we will still return a `nil` value, but [will avoid making the call](https://github.com/department-of-veterans-affairs/caseflow/pull/16403/files#diff-a5e27a49cf3fe80cb3770494bcba425adb7cf58295545d808a108bf10ffaadffR122) to `TimezoneService.address_to_timezone()` if the address does not have a country, zip code, and state. We will stop seeing these errors, but these errors are not actionable since the addresses are coming from BGS and developers do not attempt to correct those addresses in BGS. 

### 2️⃣ Addresses with a `nil` country field

Some `LegacyAppeal`s (for example, 615504, 896697, and 1110864) have `appellant_address` with all fields and values present except for a `nil` value for the country field:
```ruby
{
  :address_line_1=>"1200 Mike Fahey St",
  :address_line_2=>nil,
  :address_line_3=>nil,
  :city=>"Omaha",
  :state=>"NE",
  :country=>nil,
  :zip=>"68102"
}
``` 

Behaviour before this change | After
--- | ---
As written, the code creates an `Address` object for addresses like these and calls `TimezoneService.address_to_timezone()` which [fails when trying to determine the country](https://github.com/department-of-veterans-affairs/caseflow/blob/8e547951af3d4ff727fac940aead741f55f5b796/app/services/timezone_service.rb#L30) from the input `Address` and sends an `InvalidCountryNameError` above to Sentry and returns a `nil` value to the serializer. | With this change we are using the presence of a zip code as an indication that the country is the "US" and [creating a new `Address` object](https://github.com/department-of-veterans-affairs/caseflow/pull/16403/files#diff-a5e27a49cf3fe80cb3770494bcba425adb7cf58295545d808a108bf10ffaadffR124) before we make the call to `TimezoneService.address_to_timezone()`. If the zip code is not a valid US address that method will raise an error (as it does now), but if the zip code is valid then we will return the timezone identifier instead of `nil` value as we did before.