Add QueueTable component to explain#show (#16339)
### Description
Adds a QueueTable component to the Explain#Show route for better data display.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
Uncomment `binding.pry` in the `spec/fixes/investigate_duplicate_judge_assign_task_spec.rb` or `explain_spec.rb` RSpec and click on the "Narrative table" link in the browser tabs with the `explain` page. 
Test the filters for the Category and Type columns. Under the 'Details' column, click on any of the numbers to expand the attributes (the number indicates the number of attributes).