route send email error to #appeals-hearings (#16477)
Correctly routes VirtualHearings::SendEmail::RecipientIsDeceasedVeteran errors to the #appeals-hearings channel instead of #appeals-app-alerts ([example](https://dsva.slack.com/archives/C4JECDLSE/p1626085947132900))

The `application` parameter tells the sentry-to-slack lambda which channel to route the error to. I also moved the capture to `send_email` to lower the complexity of `perform`.