ETLBuilderJob: send Slack message on error (#16563)
Resolves [ETLBuilderJob not sending errors to Slack](https://dsva.slack.com/archives/C4JECDLSE/p1627568229201400?thread_ts=1627510303.194800&cid=C4JECDLSE)

### Description
When an error occurs in ETLBuilderJob, Sentry alerts are logged by https://github.com/department-of-veterans-affairs/caseflow/blob/c7b8a7b398b9aeb8e5a79317f0b1a7d375e69348/app/jobs/application_job.rb#L32
but never sent to Slack.

This PR logs to Slack as well, in the default `#appeals-job-alerts` channel and the `#appeals-data-workgroup` channel. Also, we don't need this job to retry, so don't re-throw the error, which will cause Shoryuken to retry the job.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
Keep an eye for ETL errors in Sentry and make sure they are sent to Slack as well.