Remove CreateEstablishClaim service object (#16557)
(This is not a sprint story. It is related to [CASEFLOW-2024](https://vajira.max.gov/browse/CASEFLOW-2024) in that it's removing untested code, but it's just a general cleanup task.)

### Description
This code:
* Has not been touched in more than 2 years (not an indictment on its own), and
* Has zero test coverage, and,
* Is never called anywhere in Caseflow (which is why its lack of test coverage went unnoticed), and
* Is not referenced in the caseflow or appeals-deployment wikis as something that may be manually used.

Pushing this PR to spark a discussion&mdash;I'd love to get consensus that it's not used at all before merging. (In theory we can bring it back from git history if we decide it's needed, but working backwards from "I think we used to have code to do this" may be tricky.)

### Acceptance Criteria
- [ ] We're comfortable with the fact that nothing uses this

### Testing Plan
🤷 Good question!