Fix bug where form being invalid on relationship change (#16590)
Resolves #1788

### Description
Fix bug with form not being valid preventing submission
This a temporary quick fix to address UAT. I'm going to dig into the issue a little more which is: The form is invalid when the user switches from Individual to Other - Individual, however, the validation process does not give any errors so it's unclear what's wrong. So adding this check will allow the form to still be submitted as long as there are no errors in the errors object.