Explain page: use Organizations to allow additional access (#16612)
Enables easy management (via the UI) of who is able to access the `explain` page.

### Description
Add `BoardProductOwners` and `CaseflowSupport` organizations, whose members will be able to access the `explain` page.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] 2 new orgs show in the Teams Management page and users can be added to those orgs
- [ ] Users of the 2 new orgs can access the `explain` page
- [ ] Reminder to add users from the `Bva` org to the appropriate new org in prod.

### Testing Plan
(Not necessary, but if you want to see org creation in action...) In Rails console, ensure the new orgs exist:
```ruby
CaseflowSupport.singleton
BoardProductOwners.singleton
```

Log in as `TEAM_ADMIN` and go to http://localhost:3000/team_management. Scroll down to:
![image](https://user-images.githubusercontent.com/55255674/129112007-28f413a2-88f2-488b-8050-a29a3a4bbd7e.png)
and add a user to any of the 2 orgs -- remember this user. Check that the user was added, e.g., `CaseflowSupport.singleton.users`.

Log in as the user you added and go to the explain page for any appeal, e.g., http://localhost:3000/explain/appeals/12bb84ff-65fb-4422-bee4-fe7553fdf5c3
![image](https://user-images.githubusercontent.com/55255674/129112593-6961b946-f268-401c-92f9-aaecb96aea66.png)

Log in as any user who is not in either of the 2 new orgs and `user.admin? => false` . Reload the page. You should see:
![image](https://user-images.githubusercontent.com/55255674/129112563-938c3a85-34a6-4100-a0b9-ef36d889d3a9.png)