Tests for CASEFLOW-1501 (#16588)
### Description
This was kind of an odd story.

[CASEFLOW-1501](https://vajira.max.gov/browse/CASEFLOW-1501) asks us to implement the backend logic from 1499's UI. It turns out that it just works™.

The main thing in this PR is a slightly-dirty feature test I added to make it easy for me to test the various combinations. It has no value as a new automated test, but I'm committing it (set to be skipped) because it could be useful later.

I then went to incorporate some of that into unit tests, but found that there wasn't really anything worthwhile to add. I'm open to adding all the combinations in the feature test, but having already written 180 lines of basically-valueless test code, I was reluctant to port it when we're already testing multiple combinations there.

So there's nothing to release here.

### Acceptance Criteria
Lifted from the Jira ticket:
* the user selects confirm on the review and confirm page 
  * Verify a new appeal stream is created that maintains the original appeals docket number and status (ex. AOD)
  * the selected tasks are created and assigned to the appropriate organizations 
    * Hearing tasks that are selected are created along with their parent/child tasks
    * If non-colocated IHP task selected verify the IHP task is assigned to substitute's POA
    * if evidence submission window is selected verify the timed hold ends on the calculated date in [~~CASEFLOW-1082~~](https://vajira.max.gov/browse/CASEFLOW-1082 "Add Review & Confirm Page for Substitution")
  * all typical parent tasks are created or are in existence for the user-selected task
  * verify tasks from source appeal are displayed in the case timeline
  * the granted substitution date is added to the about the appellant section
  * the user is taken to the Case Details page
  * verify the user sees a success banner on the Case Details page
* verify that the new task tree has been properly created

### Testing Plan
If you want, you can uncomment the `binding.pry` and test combinations by yourself. Or, run the test manually locally (overriding the skip), and see that the created task trees are reasonable.