CASEFLOW-1985 | Test coverage for JudgeTeam#ama_only_request (#16669)
_This is https://github.com/department-of-veterans-affairs/caseflow/pull/16659 with just my change. The PR it's based on was automatically squashed-and-merged so it became a mess._

### Description
Completes [CASEFLOW-1985](https://vajira.max.gov/browse/CASEFLOW-1985).

This is the second sprint story in a row where the outcome has been, "Oh wow, the requested functionality already works, but needs a test." In this case, Bar nailed it in her PR for 1982.

~This is a stacked PR building on top of https://github.com/department-of-veterans-affairs/caseflow/pull/16652~

tl;dr, we learned that the requests in 1982 and 1985 weren't quite the same: the request wasn't for *a* toggle for AMA-only on a JudgeTeam, it was for *two* toggles, one toggling whether priority push jobs for that team were AMA-only, and one toggling whether requests for more cases would be AMA-only. I want to note for posterity that, although this felt a bit more complicated, Design and Product confirmed with stakeholders that these are, indeed, desired to be two _separate_ toggles, able to be controlled independently.

Bar implemented this as two booleans in the aforementioned PR, making this work in both cases. (Thanks, Bar!) This PR merely adds some test assertions demonstrating that it works properly, to complete the rest of the PR. The logic we captured is as follows:
* If `ama_only_request` is *not* set (i.e., it is at the default value of false), case distribution behaves as normal. (The first test example I added shows this, just to make the change in the subsequent example clearer.) Per the existing logic, distributions include mostly (but not exclusively) LegacyAppeals.
* If `ama_only_request` *is* set, only AMA cases will be distributed, *except* that any legacy cases _tied to the requesting judge_ are still included.

(The behavior is similar for `ama_only_push` instead of `ama_only_request`, but that's in Bar's PR.)

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Tests pass

### Testing Plan
This is going to be a real headache to manually test. Bar and I believe it may not be of much value here, as tests more explicitly cover it.