Send status emails to coordinator user (not original recipient) (#16725)
### Description
In the work we did to send status emails to hearing coordinators. Some confusion over email addresses fields meant that the status emails were configured to send to the original recipient of the email rather than the hearing coordinator as intended.

None of these emails have gone out yet. We caught this before it caused a problem.

### Acceptance Criteria
- [x] Tests run
- [x] Tests look at the correct things
- [x] The sent_by user is in fact who we should be sending the emails to