BusinessLineReporter: Partially eager-load queries; add basic test (#16738)
### Description
I noticed that DecisionReviewsController#index was pretty slow in New Relic: https://onenr.io/0YBR635N1jO

While the slowest component was BGS calls, it also showed an average of over 4,000 "Postgres RequestIssue find" calls. This PR _partially_ improves that, but there are still some one-off queries that are less easily fixed. The focus of this innovation project was to find easy wins, so I'd like to call this PR complete: it reduces the number of queries triggered from within a loop.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Code appears safe

### Testing Plan
No behavior change. Also, I added a test for something that previously had none. 😛