Use eager loading for travel board cases in FetchHearingLocationsForVeteransJob (#16743)
### Description
Looking at Shoryuken jobs, FetchHearingLocationsForVeteranJob is extremely busy, and spends a huge amount of time making VACOLS calls.

![Screen Shot 2021-09-09 at 11 08 56 PM](https://user-images.githubusercontent.com/314663/132793287-b26a6a4f-f59d-4aea-89bb-de8996e8164f.png)

We were talking on Echo today about how, because VACOLS is hosted on the opposite side of the country, the simplest of queries will take about 100ms just due to latency, so trying to fetch data in batches is going to be a lot more efficient.

This one took me a little bit to figure out. The actual slow queries come in `AppealRepository#set_vacols_values`:
https://github.com/department-of-veterans-affairs/caseflow/blob/master/app/repositories/appeal_repository.rb#L179-L181

This was triggering 3 VACOLS queries. (This is, in turn, called from `build_appeal`.) The reference to `case_issues` is on the last line of `set_vacols_values`.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
Should be covered by unit tests.