Remove Unrecognized Appellant Appeals from StuckAppealsChecker Alert (#16759)
### Description
During the PI, I focused on reducing the number of appeals that show up in the StuckAppealsCheckerAlert.  Four appeals that currently show up in the alert have unrecognized appellants.  It was intentional to not have a dispatch task for these appeals, because we didn't want the board to send out a letter.  However, at some point in the future a bva dispatch task will need to be completed.  To ensure that the bva dispatch is completed at a later date, I created a `BvaDispatchTask` and then made it have a child `EngineeringTask`, which blocks the parent `BvaDispatchTask`.  

As a side effect of this change, these four appeals will not be selected for quality review.

I want to note one part of the rspec test which differs from my planned actions in prod.  `BvaDispatchTask.create_from_root_task(appeal.root_task)` initially broke because the assignee pool associated with the bva dispatch organization was empty.  The assignee pool was empty because it consists of non-admin bva dispatch users, and the imported appeals didn't include any such users.  As a result, I added a non-admin BVA dispatch user to each appeal json file.  This error would not occur in prod because prod data will have non-admin bva dispatch users.

### Testing Plan
1. Run the feature test. Behold!