Switch logic in fetching vacols RO information (#16766)
Resolves https://github.com/department-of-veterans-affairs/dsva-vacols/issues/220

### Description
More often than not `vacols_record.hearing_venue` is nil, thus returning `vacols_record.bfregoff` as the RO key. For the cases that are causing problems as identified in the GH issue above, `hearing_venue` is populated but as either "D2" or "5" which do not correspond to any regional office we have in our records. In these cases, `bfregoff` is still populated with an actual RO key.