Remove Unused Karma Tests from DateSelector and Refactor Regex (#16755)
Relates: #[16462](https://github.com/department-of-veterans-affairs/caseflow/pull/16462)

This PR removes file of commented out Karma tests and the unused Regex based functionality that they used to cover.

### Description
I'm slowly working on allowing us to remove Karma from Caseflow. This moves us one step closer to that by removing the Karma tests for `DateSelector`.

While reviewing those commented tests to see if I needed to port any to Jest, I found that most of the (complex, old) logic in DateSelector does not seem to be used anymore. I removed that logic, which is a big win for future developers seeking to understand this component.

I did consider removing DateSelector entirely since it is now adding very little (nothing?) to the TextField component, but I'm already extending my scope and that's one step too many for this PR.

### Testing Plan
- Search the codebase for any <DateSelector /> with a type that is not 'date'. I couldn't find any, but I'd love additional 👀
- Verify that the snapshot change for `EditAppellantInformation.test.js.snap` looks ok. I'm not certain why that changed.
- I think it's okay that the (newVal = '') call is no longer present?