Explain page: add task versions (PaperTrail) (#16822)
Resolves the [need for non-engineers to see task versions](https://dsva.slack.com/archives/CQTDX9BF0/p1633098207077700?thread_ts=1633095910.075200&cid=CQTDX9BF0)

### Description
Makes task versions accessible to non-engineers by adding a "Task versions" section to the [Explain page](https://github.com/department-of-veterans-affairs/caseflow/wiki/Explain-page-for-Appeals).

This is a really basic UI to make the data available. We can incorporate the task versions into the "Appeal Narrative" down the road if desired.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] new "Task versions" section on the Explain page

### Testing Plan
Insert a `binding.pry` into the RSpec changes.