Bug Fix - Unidentifed Issues with no decision date preventing claim establishment (#16850)
Resolves [INC19899914](https://dsva.slack.com/archives/CHX8FMP28/p1634144343149500)

### Description
As part of the recent [ITF Supplemental Claims](https://vajira.max.gov/browse/CASEFLOW-142) release, when submitting a Compensation Supplemental Claim, we check the date of the decision being contested on each active, rating request issue being submitted as part of the claim. This is done to determine which EP code will be used during claim establishment. Currently, unidentifed issues fall under the `rating` method and these issues do not require a decision date when being intaken. The update here is to exclude those unidentifed issues when checking the decision date on all active, rating request issues being submitted as part of a Compensation Supplemental claim.