Better Handling of Org Tasks w/ Cancelled User Task (#16851)
Relates to CASEFLOW-1125 and CASEFLOW-1549

### Description
After doing the prod update for CASEFLOW-1549, we realized that the default behavior of `Task.hide_from_task_snapshot` was that it would not show org tasks in the "Currently active tasks" section if they had child user tasks of the same type, regardless of the status of the child task(s). This meant that the org task could not be acted upon (for assigning to a user). This updates that logic so that open org tasks can be assigned from Case Details page even if they have a cancelled user task under them.

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Org tasks w/ status `assigned` can be assigned to a user, even if there exists an existing `cancelled` user task under them

### Testing Plan
1. Download [the exported JSON](https://github.com/department-of-veterans-affairs/caseflow/blob/yoom/appeal-188663/spec/records/appeal-188663.json) file of the example from prod
2. [Import that JSON](https://github.com/department-of-veterans-affairs/caseflow/wiki/Exporting-and-Importing-Appeals#importing) into your local Rails console (NOTE: if you see an error re `Veteran` record not existing, create a `Veteran` record and update the id to match `1140531`)
3. Check task tree, verify that `IhpColocatedTask` org task is `assigned`, user task is `cancelled`
4. Update the `assigned_to` for the org task to the default `Colocated` org (using imported one causes testing difficulties):
```ruby
org_task = Task.find(2002053616)
org_task.update!(assigned_to: Colocated.singleton)
```
5. Verify that as `VLJ_SUPPORT_ADMIN` user, you can see the org task in "Currently active tasks" and take actions, including assigning to a new user