CASEFLOW-2288: Improve Automatic Case Distribution tests (#16823)
### Description
[CASEFLOW-2288](https://vajira.max.gov/browse/CASEFLOW-2288): improves tests around Automatic Case Distribution.

There's lots more I would like to do here, but we had to somewhat timebox this rather than take a PI to rewrite all these tests.

Do not try viewing the "diff" in `distribution_spec.rb`. I moved the existing one to `distribution_spec_deprecated.rb` and created a new one named `distribution_spec.rb`, so they have nothing in common, but GitHub is still trying to show a diff. The long-term goal is to eliminate `distribution_spec_deprecated.rb`, but for now it's a useful end-to-end test.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
1. Test fixes only. Make sure tests still pass.