Request Type Dropdown Bugfix (#16876)
There is a bug in the request type dropdown (refactored last sprint) that is blocking the ability to schedule a travel board hearing. This fixes that bug by correcting the options available in the dropdown for travel board (and central) dockets.

This PR also makes an unrelated one line change in `DetailsHeader.jsx` to activate a link for Travel board hearing days.