Add wizard for fixing SSN/filenumber mismatch issue (#16842)
### Description
This PR introduces a new style of backend code I've been calling a "wizard," as a throwback to when animated paper clips would interactively walk us through multi-step processes that required human judgment.

In this case, the wizard assists Bat Team engineers in resolving the SSN/Filenumber mismatch issue. The current fix for this issue is described in an extremely long code (and not particularly self-explanatory) [snippet on the Bat Team Quickref](https://github.com/department-of-veterans-affairs/appeals-deployment/wiki/Bat-Team-Quick-Ref#ssn-stored-as-file-number).

As of yet, I'm not completely decided that a wizard is the long-term solution we want to embrace, but it has made resolution of the SSN/Filenumber mismatch issue much more pleasant. If we do want to build more wizards, certain aspects of this one are good for reüse:
- Class method that's callable with various keyword arguments
- The `#get_input` method which handles keyboard input (in a style similar to `git add -p`)

### Usage
0. If you're not currently on this branch, copy the contents of `fix_file_number_wizard.rb` into `lib/helpers/`
1. When you start up a Rails console, it should have automatically loaded the FixFileNumberWizard class. Initiating the wizard is as simple as

```ruby
> FixFileNumberWizard.run
```

### Testing Plan
- (How) should we write tests for code in `lib/helpers/`?