Improve legacy docket model (#16924)
Resolves Jira ticket [2290: Improve Legacy Docket Model](https://vajira.max.gov/browse/CASEFLOW-2290).

### Description
The goal of this PR was to improve the legacy docket model, with a focus on consolidating the distribution levers.  @jcq and I agreed to consolidate the levers in `constants/DISTRIBUTION.json`.
I also fixed a bug in the method `should_distribute?`, which was previously known as `really_distribute`, and refactored the associated tests.
I discussed with fellow engineers potentially refactoring associated logs.  We have discussion notes [here](https://hackmd.io/HaEPevsuRdaHEN6WWpRrww).  I didn't make that change in this PR because the bug fix and test update for `should_distribute?` took longer than anticipated.  I do think it would be a good refactor for the future.  

### Acceptance Criteria
- [ ] Code compiles correctly
- [ ] Automated tests pass

### Testing Plan
1. There is not a manual testing plan for this PR.  I did not introduce new functionality, and the automated tests cover the refactors that I did.