libbey/caseflow 2224 contested claims badge in docket (#16940)
Resolves [CASEFLOW-2224](https://vajira.max.gov/browse/CASEFLOW-2224)

### Description
Displays the "CC" (Contested Claim) badge in the daily docket for both AMA and Legacy hearings. 

### Testing Plan
The easiest way to test is with React Dev Tools, by finding a ContestedClaimBadge component in the Docket and toggling its `appeal.contestedClaim` prop to `true`.  

Otherwise, for AMA hearings, you would need to create or update a request issue on an appeal to have `nonrating_issue_category` of "Contested Claims - Insurance" (or one of the other categories that result in a contested claim designation). For Legacy hearings, you would need to add or update a VACOLS::Representative with the `repkey` matching `vacols_id` on the legacy appeal of interest, and give that representative a `reptype` of "C" (or one of the other types that result in a contested claim designation). 


### User Facing Changes
 - [ ] Screenshots of UI changes added to PR & Original Issue

 BEFORE|AFTER

![Screen Shot 2021-11-12 at 3 32 23 PM](https://user-images.githubusercontent.com/92495948/141595771-0bdbe050-52c5-47ec-9cd9-872e5c0fb375.png) |  ![Screen Shot 2021-11-12 at 12 20 19 PM](https://user-images.githubusercontent.com/92495948/141595781-80e1efb6-2851-4aa0-a275-43dc008a6b64.png)


![Screen Shot 2021-11-12 at 3 33 01 PM](https://user-images.githubusercontent.com/92495948/141595787-bb3d05e2-74cd-411d-b330-1844da55952d.png) | ![Screen Shot 2021-11-12 at 12 20 10 PM](https://user-images.githubusercontent.com/92495948/141595794-c6b5bd46-e182-459e-8315-f353be2c4d76.png)