Don't use veteran's SSN for VACOLS appellant if appellant is not veteran (#16976)
Resolves [INC19452345](https://dsva.slack.com/archives/CHX8FMP28/p1631043547190300)

### Description
This does not use the veteran to create a "person for appellant" for a legacy appeal, if the appellant is not the veteran, but the veteran's SSN is stored in the CORRES table.  This allows the address to be fetched from the CORRES table directly in these cases.

Re-creates https://github.com/department-of-veterans-affairs/caseflow/pull/16737 since I accidentally pulled in a TON of commits when trying to update the tests.