Bug fix for IhpDaysWaitingTooltip and ContestedClaimBadge tooltip (#16979)
### Description
When working on CASEFLOW-2593, I noticed this bug in the Transcription Team's Unassigned tab table in Queue. If you look in the gif below, you can see a Task ID I temporarily added to the top of the tooltip, which should be different for each row, but it remains the same (2001816737), always displaying the value for the last row in the table. 
![TranscriptionQueueTooltip_bug](https://user-images.githubusercontent.com/92495948/142682485-62705f62-1e63-4b9f-8ce0-5209ddb74153.gif)  

Similarly, the ContestedClaimBadge tooltip shows the text for a HearingBadge (which appears in the last row in the table).  
![CCBadge_bug](https://user-images.githubusercontent.com/92495948/142925253-79cd63b0-565a-42a3-be04-4dbd52ffea1c.gif)


To fix it, this PR adds a unique id to the tooltips, and then each row gets the correct value displayed:  
![TranscriptionQueueTooltip_fix](https://user-images.githubusercontent.com/92495948/142682493-27de2049-dea0-42ac-ac27-cef49ba02b12.gif)  

ContestedClaimBadge fixed:  
![CCBadge_fix](https://user-images.githubusercontent.com/92495948/142925317-3cdb0368-bbe9-4b39-9613-c0f84c88a536.gif)


### Testing Plan
1. View a table that shows the "Days Waiting" column. 
2. You may need to comment out lines 24-26 in client/app/queue/components/IhpDaysWaitingTooltip.jsx to see the tooltips (I didn't have any local data with values for `requestedAt`.)
3. If needed, temporarily add a unique id to the tooltip to verify that each row gets the correct value displayed.